# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Vita Track AI",
        "title": "Vita Track AI API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/documents/infiniteScroll": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Document"
                ],
                "summary": "Get Infinite Scroll Documents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Base64 encoded cursor from previous response",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of documents to fetch (default: 12)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.InfiniteScrollResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/documents/update/{id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Document"
                ],
                "summary": "Update Document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Category [Prescription or Medical Report]",
                        "name": "category",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "DocumentName",
                        "name": "document_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "DocumentDate (YYYY-MM-DD)",
                        "name": "document_date",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "tags",
                        "name": "Tags",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Document": {
            "type": "object",
            "properties": {
                "analysis_generated": {
                    "type": "boolean"
                },
                "category": {
                    "type": "string"
                },
                "document_date": {
                    "type": "string"
                },
                "document_name": {
                    "type": "string"
                },
                "file": {
                    "$ref": "#/definitions/models.File"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "description": "JSON string",
                    "type": "string"
                },
                "user_id": {
                    "description": "ID         string    `json:\"id\" gorm:\"type:uuid;default:gen_random_uuid();primaryKey\"`",
                    "type": "integer"
                }
            }
        },
        "models.File": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "fileSize": {
                    "type": "integer",
                    "format": "int64"
                },
                "id": {
                    "type": "string"
                },
                "mimeType": {
                    "type": "string"
                },
                "originalName": {
                    "type": "string"
                },
                "s3Key": {
                    "type": "string"
                },
                "storedName": {
                    "type": "string"
                },
                "uploadedBy": {
                    "type": "integer",
                    "format": "int64"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                }
            }
        },
...
        "models.InfiniteScrollResponse": {
            "type": "object",
            "properties": {
                "cursor": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Document"
                    }
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "age": {
                    "type": "integer"
                },
                "created_at": {
                    "description": "YYYY-MM-DD HH:MM:SS.microseconds stored in DB",
                    "type": "string"
                },
                "dob": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "google_id": {
                    "type": "string"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "profile_pic": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}